updated text

+++ code/ComplexPlane.cpp
@@ -84,7 +84,7 @@ void ComplexPlane::updateRender()
             {
                 m_vArray[j + i * m_pixel_size.x].position = {(float)j,(float)i};
                 ComplexPlane::mapPixelToCoords(m_pixel_size);
-                size_t iterCount = ComplexPlane::countIterations(m_plane_size);
+                int iterCount = ComplexPlane::countIterations(m_plane_size);
                 sf::Uint8 r, g, b;
                 ComplexPlane::iterationsToRGB(iterCount, r, g, b);
                 m_vArray[j + i * m_pixel_size.x].color = { r,g,b };
